Move circuit_breakers to IRBaseMapping
